requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-groq
langchain-text-splitters
langgraph
chromadb
python-dotenv
streamlit
pypdf
google-generativeai 
sentence-transformers
langchain-huggingface
fastapi
uvicorn[standard]
cryptography
pymysql
sqlalchemy
strawberry-graphql[fastapi]
langsmith
openai